add levels

+++ bot.py
@@ -2,6 +2,7 @@ import telebot, os, time, shutil, json
 import introduction.TR_to_gpt as tr
 from dotenv import load_dotenv
 from hydromet_stations.report_template import doc_report, station_map, table_map_paste
+from levels.transformation import GB_to_Profiles_tranformation
 
 # Подгрузка ключей
 load_dotenv()
@@ -22,6 +23,7 @@ def start(message : telebot.types.Message):
 - /start - основная информация о боте;
 - /introduction - создание введения;
 - /hydromet_stations - предоставление гидрометеорологической изученности по координатам.
+- /gb_to_profiles_transform - трансформация профилей global mapper в профили Profiles.
     '''.format(message.from_user.first_name)
     
      # ответ с коммандами
@@ -29,7 +31,8 @@ def start(message : telebot.types.Message):
     start_item = telebot.types.KeyboardButton("/start")
     introdiction_item = telebot.types.KeyboardButton("/introduction")
     hydromet_stations_item = telebot.types.KeyboardButton("/hydromet_stations")
-    markup.add(start_item, introdiction_item, hydromet_stations_item)
+    gb_to_profiles_transform = telebot.types.KeyboardButton("/gb_to_profiles_transform")
+    markup.add(start_item, introdiction_item, hydromet_stations_item, gb_to_profiles_transform)
 
     bot.send_message(message.chat.id, start_text, reply_markup=markup)
 
@@ -46,7 +49,7 @@ def introduction_start(message : telebot.types.Message):
 # Исполнение функции
 def introduction(message : telebot.types.Message):
     try:
-        # Проверяем статус пользователя
+        # Проверяем статус пользователя (introduction)
         if message.chat.id in user_status and user_status[message.chat.id] == 'waiting_for_TRdocument':
 
             # Получаем информацию о файле
@@ -106,16 +109,51 @@ def introduction(message : telebot.types.Message):
             # Сбрасываем статус ожидания для данного пользователя
             del user_status[message.chat.id]
         
+        # Проверяем статус пользователя (GB_transform)
+        elif message.chat.id in user_status and user_status[message.chat.id] == "waiting_for_gb_profile_zip":
+            # Получаем информацию о файле
+            file_info = bot.get_file(message.document.file_id)
+            file_path = file_info.file_path
+
+            # Скачиваем файл    
+            downloaded_file = bot.download_file(file_path)
+
+            # Сохраняем файл
+            path = "levels"
+            file_name = message.document.file_name
+            file_path = os.path.join(path, str(message.chat.id) +".zip")
+            with open(file_path, 'wb') as f:
+                f.write(downloaded_file)
+
+            bot.reply_to(message, "Обработка... ⏳")
+
+            output_zip = f"levels/{message.chat.id}_output.zip"
+            GB_to_Profiles_tranformation(str(message.chat.id), "levels", file_path, output_zip)
+
+            bot.reply_to(message, "zip-архив с файлами для Profiles:")
+            # Отправка файла клиенту
+            with open(output_zip, 'rb') as f:
+                bot.send_document(message.chat.id, f)
+            
+            # Удаление output zip
+            if os.path.exists(output_zip):
+                os.remove(output_zip)
+                print(f'Архив "{output_zip}" успешно удален.')
+
+            del user_status[message.chat.id]
+
         else:
-            bot.reply_to(message, "Пожалуйста, отправьте документ после ввода команды")
+            bot.reply_to(message, "Пожалуйста, отправьте документ после ввода команды или пришлите верный формат файла")
     
     except Exception as e:
         bot.reply_to(message, e)
         # Очистка папок
-        shutil.rmtree("introduction/input")
-        shutil.rmtree("introduction/output")
-        print(e)
+        if os.path.exists("introduction/input"):
+            shutil.rmtree("introduction/input")
 
+        if os.path.exists("introduction/output"):
+            shutil.rmtree("introduction/output")
+        print(e)
 
 # Функция изученности
 @bot.message_handler(commands=['hydromet_stations'])
@@ -125,6 +163,13 @@ def introduction_start(message : telebot.types.Message):
     bot.send_message(message.chat.id, start_text)
 
     user_status[message.chat.id] = 'waiting_for_coordinates'
+
+# Функция трансформации из GB профиля в формат profiles
+@bot.message_handler(commands=['gb_to_profiles_transform'])
+def GB_to_profiles_transform_start(message : telebot.types.Message):
+    user_status[message.chat.id] = "waiting_for_gb_profile_zip"
+    text = """Отправь zip-архив содержащий профили, сделанные в Global Mapper в формате .csv. В ответ я пришлю zip-архив с txt файлами пригодными для открытия в программе Profiles"""
+    bot.send_message(message.chat.id, text)
 
 @bot.message_handler(content_types=['text'])
 # Выполнение функции
